requirements.txt
requests
pytz
pycryptodome